Slighly rework Umbra script, make interrupt handling more elegant
The main problem in the previous implementation was the use of "bash -c
'...'" *with a redirect of stdout and stderr*.

The use of "bash -c '...'" itself is already bad practice as it e.g.
does not forward interrupts. However, this can be worked around by
adding --init to docker run. Still, when redirecting stdout, this fails
-- so it's best to avoid this redirect and use "docker logs" for
inspecting the output of the entrypoint application if necessary.